tests/requirements.txt
Faker==37.1.0
pytest==8.3.5
pytest-repeat==0.9.4
python-dotenv==1.1.0
allure-pytest
sqlalchemy
psycopg2-binary
sqlmodel
Requests==2.32.3
selenium==4.31.0
pipreqs
requests-toolbelt